+++ lib/redis.rb
@@ -2749,6 +2749,16 @@ class Redis
     self.class.new(@options)
   end
 
+  def connection
+    {
+      host:     @original_client.host,
+      port:     @original_client.port,
+      db:       @original_client.db,
+      id:       @original_client.id,
+      location: @original_client.location
+    }
+  end
+
   def method_missing(command, *args)
     synchronize do |client|
       client.call([command] + args)